leverage templates for webhook messages
This change replaces the `fmt.Sprintf` strategy for rendering messages with Go's `text/template` package, extracting sub-templates for common components. A few minor changes were made along the way:

* "Issue ... closed by" now links the repo vs just wrapping it with brackets
* (force-)pushed now pluralizes "commits" only when necessary

In addition, whenever rendering a GitHub username, if there is a linked Mattermost user, the rendered message replaces this with an @username mention for the corresponding Mattermost user.