Added code previews from permalink
For every message we now hook into the MessageWillBePosted event, and check
if it contains a permalink to a Github file.

If yes, we parse out the link, and make an API call to the Github contents API.
Get the file contents and filter out the required lines given in the permalink anchor.

Then we construct a code block in markdown and replace the original string.

Added a config key to gate this feature.