[GH-313] Add slash command to create Github Issue
Adds a new slash command to open the CreateIssue modal without needing to interact with a post.
Now we're able to create an issue without coming from a post we should ensure the server is able to handle issues being created with just a channelId rather than a postId.
If issues are created without a postId then don't append mattermost "created from" prompts to the Github issue