Use websockets to open create issue dialog
Instead of registering a hook on the front end to listen for particular messages and intercept some before they can be sent to the server, instead have the server send a websocket event back to the webapp when we need to open the create issue modal.
Set the arguments to `/github issue create` as the issue title rather than the issue description.